feat: rename `Root` to `index` in `content-runtime`
